Moved router logic content into RouteManager.
